requirements.txt
numpy>=1.21.0
scipy>=1.7.0
networkx>=2.6.0
matplotlib>=3.4.0
PyYAML>=6.0
pydantic>=2.0.0
tqdm>=4.65.0
seaborn>=0.11.0
pandas>=1.3.0
h5py>=3.6.0
colorlog>=6.0.0
ipywidgets>=8.0.0